Feat: 계정 관련 api 연결 (목) - 유저 추가 삭제 시 api 호출 및 재탐색 하도록 설정 - 유저 추가 시 계정 탐색하도록 - 유저 삭제 및 추가 시 다시 유저 탐색하도록 훅 제작

+++ apps/electron-client/src/renderer/pages/settings/components/SettingsMailList.tsx
@@ -2,6 +2,8 @@ import { AuthUser } from "@/types/authType";
 
 import useAuthenticateStore from "@stores/authenticateStore";
 
+import { useDeleteAccount } from "@hooks/useGetUser";
+
 import IconButton from "@components/common/button/IconButton";
 
 import GoogleIcon from "@assets/icons/GoogleIcon";
@@ -10,8 +12,25 @@ import DeleteIcon from "@assets/icons/DeleteIcon";
 const InnerList = ({ user }: { user: AuthUser }) => {
   const { deleteAuthUser } = useAuthenticateStore();
 
-  function handleDelete() {
-    deleteAuthUser(user);
+  const { mutateAsync: deleteAccount } = useDeleteAccount();
+
+  async function handleDelete() {
+    if (!user) return;
+
+    if (!confirm(`${user.email}\n계정을 삭제하시겠습니까?`)) return;
+
+    try {
+      const data = await deleteAccount({ accountId: user.id });
+
+      if (data.success) {
+        alert(`${user.email}\n계정이 삭제되었습니다!`);
+        deleteAuthUser(user); // 임시..
+      }
+    } catch (error) {
+      console.error("Error deleting account:", error);
+      alert("계정 삭제에 실패했습니다.");
+      return;
+    }
   }
 
   return (
